fix openBrowser bug
